+++ packages/bpk-component-fieldset/index.js
@@ -1,0 +1,3 @@
+import BpkFieldset from './src/BpkFieldset';
+
+export default BpkFieldset;