+++ packages/bpk-component-form-validation/index.js
@@ -1,0 +1,3 @@
+import BpkFormValidation from './src/BpkFormValidation';
+
+export default BpkFormValidation;