+++ packages/bpk-component-select/src/BpkSelect.js
@@ -3,19 +3,22 @@ import React, { PropTypes } from 'react';
 import './bpk-select.scss';
 
 const BpkSelect = (props) => {
-  const { valid, large, docked, dockedFirst, dockedMiddle, dockedLast, className, ...rest } = props;
   const classNames = ['bpk-select'];
+  const { valid, large, docked, dockedFirst, dockedMiddle, dockedLast, className, ...rest } = props;
+
+  // Explicit check for false primitive value as undefined is
+  // treated as neither valid nor invalid
+  const isInvalid = valid === false;
 
   if (large) { classNames.push('bpk-select--large'); }
   if (docked) { classNames.push('bpk-select--docked'); }
   if (dockedFirst) { classNames.push('bpk-select--docked-first'); }
   if (dockedMiddle) { classNames.push('bpk-select--docked-middle'); }
   if (dockedLast) { classNames.push('bpk-select--docked-last'); }
-  if (valid === false) { classNames.push('bpk-select--invalid'); }
   if (className) { classNames.push(className); }
 
   return (
-    <select className={classNames.join(' ')} {...rest} />
+    <select className={classNames.join(' ')} aria-invalid={isInvalid} {...rest} />
   );
 };
 
@@ -23,7 +26,6 @@ BpkSelect.propTypes = {
   id: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
   value: PropTypes.string.isRequired,
-  onChange: PropTypes.func.isRequired,
   className: PropTypes.string,
   valid: PropTypes.bool,
   large: PropTypes.bool,